some house cleaning
