+++ src/components/forms/checkzone/styles.scss
@@ -3,8 +3,8 @@
 @use "jambonz-ui/src/styles/mixins" as ui-mixins;
 
 .checkzone {
-  padding: vars.$px02;
-  border-radius: vars.$px01;
+  padding: ui-vars.$px02;
+  border-radius: ui-vars.$px01;
   border: 2px solid #888;
   width: 100%;
   max-width: vars.$widthinput;
@@ -14,12 +14,12 @@
     align-items: center;
 
     input {
-      margin-right: vars.$px02;
+      margin-right: ui-vars.$px02;
     }
   }
 
   &__managed {
-    margin-top: vars.$px02;
+    margin-top: ui-vars.$px02;
     cursor: not-allowed;
     opacity: 0.25;
 
@@ -28,11 +28,11 @@
     }
 
     > * + * {
-      margin-top: vars.$px02;
+      margin-top: ui-vars.$px02;
     }
 
     > label + * {
-      margin-top: vars.$px01;
+      margin-top: ui-vars.$px01;
     }
 
     &.active {
