+++ src/components/grid/styles.scss
@@ -24,11 +24,11 @@
   &__row {
     @include ui-mixins.m();
     display: grid;
-    padding: vars.$px03;
+    padding: ui-vars.$px03;
     align-items: center;
 
     > div {
-      padding-right: vars.$px02;
+      padding-right: ui-vars.$px02;
     }
   }
 
@@ -66,7 +66,7 @@
         @include grid-break2() {
           grid-column: col;
           grid-row: row 2;
-          margin-top: vars.$px02;
+          margin-top: ui-vars.$px02;
         }
       }
 
@@ -81,7 +81,7 @@
         @include grid-break2() {
           grid-column: col 2 / span 2;
           grid-row: row 2;
-          margin-top: vars.$px02;
+          margin-top: ui-vars.$px02;
         }
       }
     }
