+++ src/components/index.ts
@@ -1,7 +1,8 @@
 import { Icons } from "./icons";
 import { Toast } from "./toast";
 import { Modal, ModalClose, ModalForm } from "./modal";
-import { AccessControl } from "./acl";
+import { RequireAuth } from "./require-auth";
+import { AccessControl } from "./access-control";
 import { Obscure } from "./obscure";
 import { ClipBoard } from "./clipboard";
 import { Section } from "./section";
@@ -15,6 +16,7 @@ export {
   Modal,
   ModalClose,
   ModalForm,
+  RequireAuth,
   AccessControl,
   Obscure,
   ClipBoard,
