+++ src/components/modal/styles.scss
@@ -12,35 +12,28 @@
   justify-content: center;
   align-items: center;
   z-index: vars.$zindex00;
-  padding-left: vars.$px02;
-  padding-right: vars.$px02;
+  padding-left: ui-vars.$px02;
+  padding-right: ui-vars.$px02;
+
+  /** Remove this with `subspace` feature deprecation */
   margin: 0 !important;
 
   &__stuff {
     > * + * {
-      margin-top: vars.$px02;
+      margin-top: ui-vars.$px02;
     }
   }
 
   &__box {
-    padding: vars.$px03;
+    padding: ui-vars.$px03;
     width: 100%;
     max-width: vars.$widthinput;
-    border-radius: vars.$px02;
-    box-shadow: 0px vars.$px00 vars.$px00 rgba(0, 0, 0, 0.25);
+    border-radius: ui-vars.$px02;
+    box-shadow: 0px ui-vars.$px00 ui-vars.$px00 rgba(0, 0, 0, 0.25);
     background-color: ui-vars.$white;
 
     > * + * {
-      margin-top: vars.$px03;
-    }
-  }
-
-  &__btns {
-    display: flex;
-    justify-content: flex-end;
-
-    * + * {
-      margin-left: vars.$px02;
+      margin-top: ui-vars.$px03;
     }
   }
 }
