+++ src/components/obscure/styles.scss
@@ -19,7 +19,7 @@
 
   > button[type="button"] {
     cursor: pointer;
-    margin-left: vars.$px02;
+    margin-left: ui-vars.$px02;
 
     svg {
       stroke: ui-vars.$jambonz;
