+++ src/components/spinner/styles.scss
@@ -6,8 +6,8 @@
 .lds-ring {
   display: inline-block;
   position: relative;
-  width: vars.$px04;
-  height: vars.$px04;
+  width: ui-vars.$px04;
+  height: ui-vars.$px04;
 }
 
 .lds-ring div {
