+++ src/containers/login/styles.scss
@@ -9,8 +9,8 @@
   flex-direction: column;
   justify-content: center;
   align-items: center;
-  padding-left: vars.$px02;
-  padding-right: vars.$px02;
+  padding-left: ui-vars.$px02;
+  padding-right: ui-vars.$px02;
 
   .notfound {
     text-align: center;
@@ -18,7 +18,7 @@
 
   header {
     position: fixed;
-    padding: vars.$px03;
+    padding: ui-vars.$px03;
     top: 0;
     left: 0;
     width: 100%;
@@ -30,7 +30,7 @@
   footer {
     color: ui-vars.$grey;
     position: fixed;
-    padding: vars.$px03;
+    padding: ui-vars.$px03;
     bottom: 0;
     left: 0;
     width: 100%;
@@ -52,12 +52,12 @@
   }
 
   form {
-    margin-top: vars.$px04;
+    margin-top: ui-vars.$px04;
     width: 100%;
     max-width: ui-vars.$width-small;
     background-color: ui-vars.$white;
-    padding: vars.$px03;
-    border-radius: vars.$px02;
+    padding: ui-vars.$px03;
+    border-radius: ui-vars.$px02;
     flex-direction: column;
     justify-content: center;
     align-items: center;
@@ -66,7 +66,7 @@
     color: ui-vars.$dark;
 
     > * + * {
-      margin-top: vars.$px02;
+      margin-top: ui-vars.$px02;
     }
 
     > input,
