--- src/containers/styles.scss
@@ -1,9 +1,0 @@
-@use "src/styles/vars";
-@use "jambonz-ui/src/styles/vars" as ui-vars;
-@use "jambonz-ui/src/styles/mixins" as ui-mixins;
-
-.notfound {
-  .btn {
-    margin-top: vars.$px03;
-  }
-}
