--- src/styles/_forms.scss
@@ -1,52 +1,0 @@
-@use "./vars";
-@use "jambonz-ui/src/styles/vars" as ui-vars;
-@use "jambonz-ui/src/styles/mixins" as ui-mixins;
-
-.form {
-  > * + * {
-    margin-top: vars.$px02;
-  }
-
-  label {
-    @include ui-mixins.m();
-    @include ui-mixins.font-medium();
-    display: flex;
-    align-items: center;
-
-    input[type="radio"] {
-      margin: 0 vars.$px01 0 0;
-    }
-
-    button[type="button"] {
-      margin-left: vars.$px01;
-      cursor: pointer;
-    }
-
-    svg {
-      stroke: ui-vars.$teal;
-      width: 18px;
-      height: 18px;
-    }
-  }
-
-  fieldset {
-    width: 100%;
-    display: flex;
-    justify-content: flex-start;
-    flex-direction: column;
-
-    label {
-      max-width: fit-content;
-    }
-
-    > * + * {
-      margin-top: vars.$px01;
-    }
-  }
-
-  input[type="text"],
-  input[type="password"] {
-    width: 100%;
-    max-width: vars.$widthinput;
-  }
-}
