+++ src/styles/_vars.scss
@@ -1,11 +1,5 @@
 @use "jambonz-ui/src/styles/vars" as ui-vars;
 
-$px00: 4px;
-$px01: 8px;
-$px02: 16px;
-$px03: 32px;
-$px04: 64px;
-
 $ico00: 18px;
 $ico01: 24px;
 $ico03: 36px;
